backend_v2/requirements-titiler.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
numpy>=1.21.6,<2.0
pillow==10.3.0
rasterio>=1.3.10
pyproj==3.6.1
rio-tiler>=7.0,<8.0
